codeforge/requirements.txt
langchain
langchain-openai
langchain-google-genai
openai
python-dotenv
fastapi
uvicorn 